refactor (network): separate user and network settings
