+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/data/di/DataSourcesModule.kt
@@ -4,6 +4,7 @@ import com.personalization.sdk.data.repositories.network.NetworkDataSource
 import com.personalization.sdk.data.repositories.notification.NotificationDataSource
 import com.personalization.sdk.data.repositories.preferences.PreferencesDataSource
 import com.personalization.sdk.data.repositories.recommendation.RecommendationDataSource
+import com.personalization.sdk.data.repositories.user.UserSettingsDataSource
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
@@ -17,9 +18,13 @@ class DataSourcesModule {
 
     @Provides
     @Singleton
-    fun provideNetworkDataSource(
+    fun provideNetworkDataSource() = NetworkDataSource()
+
+    @Provides
+    @Singleton
+    fun provideUserDataSource(
         preferencesDataSource: PreferencesDataSource
-    ) = NetworkDataSource(
+    ) = UserSettingsDataSource(
         preferencesDataSource = preferencesDataSource
     )
