+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/domain/repositories/UserSettingsRepository.kt
@@ -1,0 +1,12 @@
+package com.personalization.sdk.domain.repositories
+
+interface UserSettingsRepository {
+
+    fun initialize(
+        shopId: String,
+        seance: String?,
+        segment: String,
+        stream: String,
+        userAgent: String
+    )
+}
