+++ personalization-sdk/src/main/kotlin/com/personalization/sdk/domain/usecases/network/InitNetworkUseCase.kt
@@ -8,20 +8,10 @@ class InitNetworkUseCase @Inject constructor(
 ) {
 
     fun invoke(
-        baseUrl: String,
-        shopId: String,
-        seance: String?,
-        segment: String,
-        stream: String,
-        userAgent: String
+        baseUrl: String
     ) {
         networkRepository.initialize(
-            baseUrl = baseUrl,
-            shopId = shopId,
-            seance = seance,
-            segment = segment,
-            stream = stream,
-            userAgent = userAgent
+            baseUrl = baseUrl
         )
     }
 }
